+++ inputs/single_nid_mult_modif.rs
@@ -1,13 +1,13 @@
-#[atomic]
-#[nids(x)]
-fn update(x:& mut u16) -> () {
+// #[atomic]
+// #[nids(x)]
+fn update(x:& mut i32) -> () {
     *x = 8;       
     *x = 3;
     *x = 8;
 }
 
-#[atomic]
-#[nids(x)]
+// #[atomic]
+// #[nids(x)]
 fn main() {
     let mut x = 0;
     update(&mut x);
